refactor(app): extract popup/overlay/lsp-confirmation orchestrators (phase 3g)
Move three loosely-related clusters out of render.rs into a new
app/popup_overlay_actions.rs sibling:

  - Overlay management: add_overlay, remove_overlay,
    remove_overlays_in_range, clear_overlays
  - Popup lifecycle: show_popup, hide_popup, dismiss_transient_popups,
    scroll_popup, on_editor_focus_lost, clear_popups, popup_select_next,
    popup_select_prev, popup_page_down, popup_page_up
  - LSP confirmation popup: show_lsp_confirmation_popup,
    handle_lsp_confirmation_response, notify_lsp_current_file_opened,
    has_pending_lsp_confirmation

All three clusters dispatch through the buffer's Event API; none of
them are render-time drawing code.

render.rs goes from 2582 -> 2136 lines (-446). 2026 tests still pass.
