+++ crates/fresh-editor/src/app/mod.rs
@@ -36,6 +36,7 @@ mod on_save_actions;
 mod path_utils;
 mod plugin_commands;
 mod popup_actions;
+mod popup_overlay_actions;
 mod prompt_actions;
 mod recovery_actions;
 mod regex_replace;
